chore(tests): create assert odm trait
